% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_file_rename.R
\name{bulk_file_rename}
\alias{bulk_file_rename}
\title{Bulk search and replace in filenaames
Can do a single replacement if p and r given, or pr, a datatable of p and rs
Also can specify prefixes and suffixes to add}
\usage{
bulk_file_rename(
  datafolder = NULL,
  p = " ",
  r = "",
  suffix = "",
  prefix = "",
  pr = NULL
)
}
\arguments{
\item{datafolder}{the folder with files to rename}

\item{p}{pattern to search}

\item{r}{replacement}

\item{suffix}{end of filename}

\item{prefix}{start of filename}

\item{pr}{a datatable of p and rs}
}
\value{
Returns a list of data.tables for each type (task, questionnaire, continuous)
}
\description{
Bulk search and replace in filenaames
Can do a single replacement if p and r given, or pr, a datatable of p and rs
Also can specify prefixes and suffixes to add
}
