% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_c_elements.R
\name{clip_c_elements}
\alias{clip_c_elements}
\title{takes a vector of values and puts them in a list with quotes and commas to put in code}
\usage{
clip_c_elements(x, containing = NULL)
}
\description{
takes a vector of values and puts them in a list with quotes and commas to put in code
}
