% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_merger.R
\name{data_merger}
\alias{data_merger}
\title{load in and merge data from multiple files into one data.table}
\usage{
data_merger(
  ending = "txt",
  contains = "",
  d = ",",
  h = TRUE,
  datafile = NULL,
  gorilla_clean = FALSE,
  datafolder = NULL
)
}
\arguments{
\item{ending}{string at the end of the filename, ie filetype.}

\item{contains}{string to identify datafile, eg 'task'}

\item{d}{delimiter for data, default comma}

\item{h}{are there headers in data files?}

\item{datafile}{example file. if this is supplied then don't ask user}

\item{datafolder}{give the file location of a folder instead of eg file}
}
\value{
A data table of all datafiles ending in \code{ending} containing \code{contains}.
}
\description{
It asks for one of the files in the DATA folder and merges
all the files in there strung together in one dataframe
}
