% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_merger_gorilla.R
\name{data_merger_gorilla}
\alias{data_merger_gorilla}
\title{Merge and load in data from a gorilla download}
\usage{
data_merger_gorilla(datafile = NULL, ending = "csv", ...)
}
\arguments{
\item{datafile}{example file. if this is supplied then don't ask user}

\item{ending}{string at the end of the filename, ie filetype.}
}
\value{
A list of two or three data tables of gorilla data.
}
\description{
It asks for one of the files in the un zipped DATA folder and
loads in the task and the questionnaire data separately
It will try and figure out if your questionnaire data is wide or long
and collate it accordingly
Note that the functions for processing questionnaire data in this package assume long
}
