% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclorr.R
\name{exclorr}
\alias{exclorr}
\title{Explore correlations between large numbers of DVs}
\usage{
exclorr(
  data,
  var1,
  var2,
  bayes_factor = F,
  plim = 0.05,
  cplot = F,
  cplot_title = "Correlation Matrix",
  outp = "analysis"
)
}
\arguments{
\item{data}{data}

\item{var1}{list of DV names that will be rows}

\item{var2}{list of DV names that will be cols}

\item{bayes_factor}{do you want bayes analysis? banal}

\item{plim}{cut off for p-value reporting}

\item{cplot}{generate a correlation matrix plot}

\item{outp}{name of folder to save plot in, set to blank if no save needed}
}
\description{
Takes a data set and two lists of DVs
looks at correlations between one set and another
r values, p values and bayes factors
}
