% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyethinkdata-package.R
\docType{package}
\name{eyethinkdata-package}
\alias{eyethinkdata}
\alias{eyethinkdata-package}
\title{eyethinkdata: Tools for processing behavioural data}
\description{
A Toolbox for behavioural data. Processes raw downloads from gorilla, preps behavioural dats for analysis and extracts and scores information from questionnaires. Performs basic plotting of RDIs (pirate plots), and correlations
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dcr-eyethink/eyethinkdata}
  \item \url{https://dcr-eyethink.github.io/eyethinkdata/}
  \item Report bugs at \url{https://github.com/dcr-eyethink/eyethinkdata/issues}
}

}
\keyword{internal}
