% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_datalist.R
\name{filtered_datalist}
\alias{filtered_datalist}
\title{filters whole data set to values of specified column}
\usage{
filtered_datalist(data, filter_col = NULL, filter_value = NULL)
}
\arguments{
\item{data}{list of data.tables}

\item{filter_col}{save out a subset of data, based on this column}

\item{filter_value}{using this value as filter}
}
\description{
filters whole data set to values of specified column
}
