% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorilla_eye_plot.R
\name{gorilla_eye_plot}
\alias{gorilla_eye_plot}
\title{Takes one trial (or trial type) of data and optional image file an makes heat map}
\usage{
gorilla_eye_plot(
  data,
  background = NULL,
  zones = NULL,
  plot_types = c("heat", "path"),
  outp = "analysis",
  title = "plot",
  contrast = NULL
)
}
\arguments{
\item{data}{gorilla ET data}

\item{background}{for background}
}
\value{
image
}
\description{
Generates plots from imported gorilla data that has been processed by gorilla_eye_processing
}
