% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorilla_iatanalysis.R
\name{gorilla_iatanalysis}
\alias{gorilla_iatanalysis}
\title{processes IAT data from gorilla}
\usage{
gorilla_iatanalysis(
  data = NULL,
  outp = "analysis",
  exclude_pids = NULL,
  exclude_items = NULL,
  ...
)
}
\arguments{
\item{data}{data list from gorilla import, if missing, I'll ask fora folder of gorilla downloads}

\item{outp}{output foldername}

\item{exclude_pids}{get rid of these guys}

\item{exclude_items}{get rid of these guys in items column}

\item{...}{pass to mypirate for plotting}
}
\description{
Runs a full analysis on IAT data from a gorilla experiment
that is based on standard template.
You can pass arguments to mypirate that is drawing the main plot
}
