% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorilla_q_bmis.R
\name{gorilla_q_bmis}
\alias{gorilla_q_bmis}
\title{Processes BMIS mood survey}
\usage{
gorilla_q_bmis(data, full = F)
}
\arguments{
\item{data}{data list or just the data_q}
}
\description{
Processes BMIS mood survey
}
