% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorilla_q_checkbox.R
\name{gorilla_q_checkbox}
\alias{gorilla_q_checkbox}
\title{Scores checkbox items}
\usage{
gorilla_q_checkbox(data, checkkey = "problem", newcolnames = NULL)
}
\arguments{
\item{data}{data list or just the data_q}

\item{checkkey}{what key was used in QB1 for this item}
}
\description{
Takes output from a checkbox questionniare (QB1) in gorilla and returns data in a wide format with one column per ranked item
}
