% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorilla_q_keyed.R
\name{gorilla_q_keyed}
\alias{gorilla_q_keyed}
\title{Processes a gorilla QB1 questionnaire using a survey key}
\usage{
gorilla_q_keyed(
  data,
  keyout = FALSE,
  qlist = NULL,
  items = F,
  key_folder = NULL
)
}
\arguments{
\item{data}{data list or just the data_q}

\item{keyout}{generate a new key from the data}

\item{key_folder}{location of a central folder where you keep survey keys}
}
\description{
The first time this is run, set keyout to TRUE to generate a key
The key to the survey should either be in a central key_folder,
or in the working directory in a folder called survey_key
}
\details{
When you have a new key, edit it in excel or text editor,
then save it as .csv file. There is one survey key for each blob in gorilla.
There is a row for every different survey item in that blob.
They are identified by what you wrote in the 'key' box when making the survey in gorilla.
You can change the numbers in the columns sum, ScaleNam, Subscore and qual, to score/summarize your survey.
You will get an output that has one row per person, and one column that scores all the answers,
and additionally other columns for subscales or text answers.
Here's what the columns mean:
sum - how this item contributes to scoring: set to 1 to add up, 0 to ignore and -1 for reverse score
rev - if it is to be reversed scored, then subtract the answer from this number. eg assuming a 7 point scale, I've set this to 8.
ScaleName - you will end up with one row per person and a variable with this name (eg IQ) summarizing all items. You can have one or many different scales in the same questionnaire and key
Subscore  - you can break the scales down further into subscales. Name it here and it will also appear on output as a scored column (eg IQ_verbal)
qual - If this item is a non numeric or qualitative response (eg a text box) then put a 1 here. It won't be summarized but till also be reported in output in a column
ignore - If the response matches this answer, then don't use it in the scale calculation. This is used, eg, for scales where 1-7 is the answer you want to code, but people can also enter 8 for 'not applicable'.
}
