% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorilla_q_parse_qb2.R
\name{gorilla_q_parse_qb2}
\alias{gorilla_q_parse_qb2}
\title{Get subject entered, unscored data like age and gender from gorilla. Will give you the qual or quantised versions}
\usage{
gorilla_q_parse_qb2(
  data,
  qlist = NULL,
  preface = NULL,
  strip = NULL,
  pd = NULL
)
}
\arguments{
\item{data}{raw data packet from from gorilla import}

\item{qlist}{list of Task.Name, ie questionnaire blob names, that you want to process}

\item{preface}{add some text to all variables}

\item{strip}{get rid of either the quant or qual}
}
\description{
Get subject entered, unscored data like age and gender from gorilla. Will give you the qual or quantised versions
}
