% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorilla_q_performance_engagement.R
\name{gorilla_q_performance_engagement}
\alias{gorilla_q_performance_engagement}
\title{Processes narrative engagement questionnaire}
\usage{
gorilla_q_performance_engagement(data, full = FALSE, taskname = "Engagement")
}
\arguments{
\item{data}{data list or just the data_q}
}
\description{
From https://app.gorilla.sc/admin/task/352459/
 Brown & Novak (2007)
 https://wolfbrown.com/wp-content/uploads/Report-Assessing-The-Intrinsic-Impacts-of-a-Live-Performance.pdf
}
