% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorilla_q_tipi.R
\name{gorilla_q_tipi}
\alias{gorilla_q_tipi}
\title{TIPI ten item personality scale}
\usage{
gorilla_q_tipi(data, qname = c("Big 5 ten items", "Big 5 Personality", "TIPI"))
}
\arguments{
\item{data}{data list or just the data_q}

\item{q_name}{Name of the questionnaire}
}
\description{
TIPI ten item personality scale
}
