% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_assignment_files.R
\name{lab_assignment_files}
\alias{lab_assignment_files}
\title{Goes through assignment download from moodle, checks files and unzips}
\usage{
lab_assignment_files(assignmentfolder)
}
\description{
Goes through assignment download from moodle, checks files and unzips
}
