% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedmodel.R, R/output_mm.R
\name{mixedmodel}
\alias{mixedmodel}
\title{Runs a mixed model and out puts results and plots}
\usage{
mixedmodel(formula, data, outp = "analysis", ...)

mixedmodel(formula, data, outp = "analysis", ...)
}
\arguments{
\item{formula}{same as LME}

\item{outp}{folder for saving. Set blank if not required}
}
\description{
Runs an mixed model and gives a text output of means,
and stats

Turns an mixed model into a text output of means,
and stats
}
