% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_anv.R
\name{output_anv}
\alias{output_anv}
\title{Outputs result of a anova using mypirate}
\usage{
output_anv(mm, t = "", outp = "analysis", cond1 = NULL, cond2 = NULL)
}
\arguments{
\item{mm}{mixed model with two factors}

\item{t}{title for plot and text files}

\item{outp}{name of data output folder}

\item{cond1}{first condition, guessed if ommitted}

\item{cond2}{second condition, guessed if ommitted}
}
\description{
Turns an ANOVA into a text output of means,
and stats
}
\details{
NOT IMPLEMENTED YET
}
