% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_mm.R
\name{output_mm}
\alias{output_mm}
\title{Outputs result of a mixed model}
\usage{
output_mm(mm, t = "", outp = "analysis", cond1 = NULL, cond2 = NULL)
}
\arguments{
\item{mm}{mixed model with two factors}

\item{t}{title for plot and text files}

\item{outp}{name of data output folder}

\item{cond1}{first condition, guessed if ommitted}

\item{cond2}{second condition, guessed if ommitted}
}
\description{
Turns an mixed model into a text output of means,
and stats
}
