% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pid_merge.R
\name{pid_merge}
\alias{pid_merge}
\title{takes all of the datatables passed to it, and merges on pid or whetever specified in link, getting rid of duplicates}
\usage{
pid_merge(..., link = "pid")
}
\arguments{
\item{...}{all the data tables}

\item{link}{cols to link on}
}
\description{
takes all of the datatables passed to it, and merges on pid or whetever specified in link, getting rid of duplicates
}
