% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_anv.R
\name{plot22}
\alias{plot22}
\title{Outputs result of a two way ANOVA, with bar plots}
\usage{
plot22(
  anv,
  t = "",
  outp = "analysis",
  cond1 = NULL,
  cond2 = NULL,
  cols = NULL,
  pirate = F,
  barchart = F,
  maineffects = T
)
}
\arguments{
\item{anv}{anova model, two factors}

\item{t}{title for plot and text files}

\item{outp}{name of data output folder}

\item{cond1}{first condition, guessed if ommitted}

\item{cond2}{second condition, guessed if ommitted}

\item{cols}{plot colours}

\item{pirate}{produce pirate plot}

\item{barchart}{produce barchart}
}
\description{
Turns an anova model into a text output of means,
anova table and bar plot interactions
}
