% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotscat.R
\name{plotscat}
\alias{plotscat}
\title{Plot scattergram for two or three variables, optionally split by group}
\usage{
plotscat(
  data,
  x,
  y,
  z = NULL,
  cl = 2,
  condcols = NULL,
  cleand = FALSE,
  results_title = TRUE,
  title_text = NULL,
  outp = "analysis",
  w = 6,
  h = 6,
  label = NULL,
  bayes_factor = F
)
}
\arguments{
\item{data}{data frame with two cols to be correlated}

\item{x}{name of x column}

\item{y}{name of y column}

\item{z}{name of column, either numeric DV or a factor for splitting groups}

\item{results_title}{do you want all the stats in the title?}

\item{title_text}{optional text for title}

\item{outp}{name of folder to save plot in, set to blank if no save needed}

\item{bayes_factor}{do you want bayes factor calculated, needs BayesFactor package}
}
\description{
this plots x  against y if there are two;
if z is given and is a number, plots x against z as well
Or z can be is a between subjects factor, will plot x y for the groups given in z
and tests for slope differences
}
