% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotscat.R
\name{plotscat_split}
\alias{plotscat_split}
\title{Plot scattergram for two variables, split by a group}
\usage{
plotscat_split(
  data,
  x,
  y,
  z,
  cl = 2,
  condcols = NULL,
  results_title = TRUE,
  outp = "",
  bayes_factor = F
)
}
\arguments{
\item{data}{data frame with two cols to be correlated}

\item{x}{name of x column}

\item{y}{name of y column}

\item{z}{name of column, a factor splitting groups}

\item{cl}{=2 plots the sub groups, =1 plots just main.}

\item{condcols}{colours for two groups}
}
\description{
this plots x against y, split by z
and it will plot x y for the two groups
and tests for slope differences
}
