% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin_datalist.R
\name{readin_datalist}
\alias{readin_datalist}
\title{reads in a set of data files and stores them in a list}
\usage{
readin_datalist(
  folder = "processed",
  flist = NULL,
  data = list(),
  drp = F,
  token = NULL
)
}
\arguments{
\item{folder}{where to find the data, defaults to 'processed'}

\item{flist}{what files do you want from folder? Defaults to all of them}

\item{data}{already existing list to add data to}

\item{drp}{is this to be downloaded from dropbox URL?}

\item{token}{dropbox token}
}
\description{
reads in a set of data files and stores them in a list
}
