% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacecondcols.R
\name{replacecondcols}
\alias{replacecondcols}
\title{returns either a set of default colours, or if condition names are also colours, returns those}
\usage{
replacecondcols(condlevels)
}
\arguments{
\item{condlevels}{the levels of a factor in a condition column}
}
\description{
returns either a set of default colours, or if condition names are also colours, returns those
}
