% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveout_datalist.R
\name{saveout_datalist}
\alias{saveout_datalist}
\title{saves out data tables in a list to a folder containing .csv files}
\usage{
saveout_datalist(data, folder = "processed")
}
\description{
saves out data tables in a list to a folder containing .csv files
}
