% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecor.R
\name{singlecor}
\alias{singlecor}
\title{single correlation with r, p value and optionally a bayes factor}
\usage{
singlecor(data, x, y, bayes_factor = F)
}
\arguments{
\item{data}{data frame with two cols to be correlated}

\item{x}{name of x column}

\item{y}{name of y column}

\item{bayes_factor}{do you want bayes factor calculated, needs BayesFactor package}
}
\description{
single correlation with r, p value and optionally a bayes factor
}
