% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_collator_gorilla.R
\name{data_collator_gorilla}
\alias{data_collator_gorilla}
\title{Merge and load in data from multiple gorilla downloads}
\usage{
data_collator_gorilla(datafolder = NULL, ...)
}
\arguments{
\item{datafolder}{the folder with a collection of gorilla downloads}
}
\value{
Returns a list of data.tables for each type (task, questionnaire, continuous)
}
\description{
Asks for a folder that can contain multiple data folders, zipped or un zipped, downloaded from gorilla
}
