% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorilla_eye_processing.R
\name{gorilla_eye_processing}
\alias{gorilla_eye_processing}
\title{Takes all data from gorilla import,
filters for named displays and returns simplified data,}
\usage{
gorilla_eye_processing(
  data,
  td = NULL,
  link = c("pid", "sid"),
  zone = "gorilla",
  conf_threshold = 0.5
)
}
\arguments{
\item{data}{compiled  data from a gorilla import}

\item{td}{if trail data already generated, will add eye info to this}
}
\value{
eye data with simplified data
}
\description{
Takes all data from gorilla import,
filters for named displays and returns simplified data,
}
