% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorilla_q_keyed.R
\name{gorilla_q_keyed}
\alias{gorilla_q_keyed}
\title{Processes a gorilla QB1 questionnaire using a survey key}
\usage{
gorilla_q_keyed(
  data,
  keyout = FALSE,
  items = F,
  key_folder = "/Users/dcr/Dropbox/Rwork/resources/"
)
}
\arguments{
\item{data}{data list or just the data_q}

\item{keyout}{generate a new key from the data}

\item{key_folder}{location of a central folder where you keep survey keys}
}
\description{
The first time this is run, set keyout to TRUE to generate a key
The key to the survey should either be in a central key_folder,
or in the working directory in a folder called survey_key
}
