% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypirate.R
\name{mypirate}
\alias{mypirate}
\title{Outputs a pirate plot (RDI)}
\usage{
mypirate(
  data,
  colour_condition = NULL,
  x_condition = "variable",
  facet_condition = NULL,
  facet_scales = "fixed",
  pid_average = F,
  dv,
  reorder = F,
  pid = "pid",
  dodgewidth = 0.8,
  plot_condition = NULL,
  cond = NULL,
  cond2 = NULL,
  facetby = NULL,
  ylim = NULL,
  xlim = NULL,
  w = NULL,
  h = 6,
  title = NULL,
  outp = "analysis",
  cols = NULL,
  pred_line = F,
  error_bar_data = NULL,
  pred = NULL,
  pred_means = NULL,
  pred_bar = T,
  xlabs = NULL,
  xlabpos = 1.1,
  error_data = NULL,
  cflip = F,
  norm = F,
  bars = F,
  violin = T,
  dots = T,
  splitV = F,
  svw = 1,
  dot_h_jitter = 0,
  line = F,
  error_bars = T,
  useall = F,
  legend = T,
  title_overide = F,
  combine_plots = list(),
  combine_position = "right",
  elementinc = NULL,
  ...
)
}
\arguments{
\item{data}{data with one person per line, excluding rows with use=0}

\item{colour_condition}{colour split}

\item{x_condition}{x axis split (if not, specified colour condition used for x axis too)}

\item{facet_condition}{for faceting}

\item{pid_average}{plot an average of each participant's dv over the named conditions}

\item{dv}{name of single dv column, or multiple columns, in which case they will be split by x_condition unless colour or facet condition set to 'variable'}

\item{pid}{whats the name of col that identifies individuals}

\item{plot_condition}{instead of passing individually, you can give a vector of up to 3}

\item{cols}{specify the colours to use}

\item{error_data}{distribution for mean, eg from Bayes analysis, to replace SE}

\item{cflip}{flip to horizontal plot}

\item{norm}{normalise / z-score values for comparison across scales}

\item{useall}{ignore the use column and plot all rows}
}
\description{
Defaults to violin plot with error bars and dots, but elements such as dots, bars, violin, error_bars can be turned on or off
}
